requirements.txt
fastapi
uvicorn
opencv-python
numpy
torch
torchvision
qdrant-client
python-multipart
Pillow
